+++ src/core/profiles/get-paginated-profiles.service.ts
@@ -11,7 +11,6 @@ export async function fetchPaginatedProfiles(
   currentSeed: number | null
 ): Promise<Response> {
   try {
-    console.log(page, limit, currentSeed);
     const response = await axios.get(
       `${URL}?page=${page}&limit=${limit}&random=${currentSeed}`
     );
