+++ src/features/profile/profile-content/use-download-curriculum-pdf.tsx
@@ -16,8 +16,7 @@ const useDownloadCurriculumPDF = (profile: Profile) => {
       if (!profile) return;
       setLoading(true);
       const response = await exportCurriculumPDF(profile.userId);
-
-      console.log(loader);
+      
       const file = new Blob([response.data], { type: "application/pdf" });
       const fileURL = URL.createObjectURL(file);
       const link = document.createElement("a");
